+++ tests/sinks/test_ndjson.py
@@ -6,7 +6,7 @@ from pydantic import UUID4
 from mex.common.models import ExtractedData
 from mex.common.settings import BaseSettings
 from mex.common.sinks.ndjson import write_ndjson
-from mex.common.types import Identifier, Timestamp
+from mex.common.types import Identifier, TemporalEntity
 
 
 class DummyEnum(Enum):
@@ -18,7 +18,7 @@ class Thing(ExtractedData):
     str_attr: str
     enum_attr: DummyEnum | None = None
     uuid_attr: UUID4 | None = None
-    ts_attr: Timestamp | None = None
+    ts_attr: TemporalEntity | None = None
 
 
 def test_write_ndjson() -> None:
@@ -31,7 +31,7 @@ def test_write_ndjson() -> None:
             identifier="3", str_attr="baz", uuid_attr=UUID(int=42, version=4)
         ),
         Thing.model_construct(
-            identifier="4", str_attr="dat", ts_attr=Timestamp(2000, 1, 1)
+            identifier="4", str_attr="dat", ts_attr=TemporalEntity(2000, 1, 1)
         ),
     ]